Replace enum-variants with strum and named_type
