--- lib/enum-variants-derive/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "enum-variants-derive"
-version = "0.1.0"
-authors = ["Jeroen van Straten"]
-edition = "2018"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-syn = "0.15"
-proc-quote = "0.2"
