--- lib/enum-variants/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "enum-variants"
-version = "0.1.0"
-authors = ["Jeroen van Straten"]
-edition = "2018"
-
-[dependencies]
-failure = "0.1"
-
-[dependencies.enum-variants-derive]
-path = "../enum-variants-derive"
