+++ src/core/common/mod.rs
@@ -3,3 +3,4 @@ pub mod error;
 pub mod log;
 pub mod protocol;
 pub mod types;
+pub mod util;
